+++ src/lib/components/Accordion/header.svelte
@@ -3,9 +3,9 @@
 </script>
 
 <script lang="ts">
-	import { useActions } from '$lib/helpers/useActions';
+	import { useActions } from '$lib/internal/helpers/useActions';
 
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 
 	type $$Props = AccordionHeaderProps;
 </script>
