+++ src/lib/components/Arrow/root.svelte
@@ -1,5 +1,5 @@
 <script lang="ts" context="module">
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 
 	export type ArrowRootProps = BaseProps<'svg'> & {
 		width?: number;
