+++ src/lib/components/Avatar/fallback.svelte
@@ -1,7 +1,7 @@
 <script lang="ts" context="module">
-	import { useActions } from '$lib/helpers/useActions';
+	import { useActions } from '$lib/internal/helpers/useActions';
 
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 	import { getAvatarRootContext } from './root.svelte';
 
 	export type AvatarFallbackProps = BaseProps<'span'> & {
