+++ src/lib/components/Collapsible/content.svelte
@@ -5,10 +5,10 @@
 </script>
 
 <script lang="ts">
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 	import { slide, type SlideParams } from 'svelte/transition';
 	import { getRootContext } from './root.svelte';
-	import { useActions } from '$lib/helpers/useActions';
+	import { useActions } from '$lib/internal/helpers/useActions';
 
 	type $$Props = CollapsibleContentProps;
