+++ src/lib/components/Collapsible/root.svelte
@@ -20,10 +20,10 @@
 </script>
 
 <script lang="ts">
-	import { reactiveContext } from '$lib/helpers/reactiveContext';
-	import { useActions } from '$lib/helpers/useActions';
+	import { reactiveContext } from '$lib/internal/helpers/reactiveContext';
+	import { useActions } from '$lib/internal/helpers/useActions';
 
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 
 	type $$Props = CollapsibleRootProps;
 	export let open = false;
