+++ src/lib/components/Collapsible/trigger.svelte
@@ -3,9 +3,9 @@
 </script>
 
 <script lang="ts">
-	import { useActions } from '$lib/helpers/useActions';
+	import { useActions } from '$lib/internal/helpers/useActions';
 
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 	import { createEventDispatcher } from 'svelte';
 	import { getRootContext } from './root.svelte';
