+++ src/lib/components/Dialog/title.svelte
@@ -1,5 +1,5 @@
 <script lang="ts" context="module">
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 	import { getDialogRootContext } from './root.svelte';
 
 	export type DialogTitleProps = BaseProps<'h2'>;
