+++ src/lib/components/Progress/root.svelte
@@ -1,7 +1,7 @@
 <script lang="ts" context="module">
-	import { reactiveContext } from '$lib/helpers/reactiveContext';
-	import { useActions } from '$lib/helpers/useActions';
-	import type { BaseProps } from '$lib/types';
+	import { reactiveContext } from '$lib/internal/helpers/reactiveContext';
+	import { useActions } from '$lib/internal/helpers/useActions';
+	import type { BaseProps } from '$lib/internal/types';
 
 	export type ProgressRootProps = BaseProps<'progress'> & {
 		/** The maximum progress value. */
