+++ src/lib/internal/components/Popper/anchor.svelte
@@ -1,5 +1,5 @@
 <script lang="ts" context="module">
-	import type { BaseProps } from '$lib/types';
+	import type { BaseProps } from '$lib/internal/types';
 	import { getRootContext } from './root.svelte';
 
 	export type PopperAnchorProps = BaseProps;
