--- src/lib/internal/index.ts
@@ -1,4 +1,0 @@
-export * from './components';
-export * from './helpers';
-export * from './actions';
-export * from './stores';
